requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
